changed home notifications commit

+++ src/app/home/home.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductService } from '../product.service';
 import { CartService } from '../cart.service';
+import { AuthService } from '../auth.service'; 
 
 @Component({
   selector: 'app-home',
@@ -9,11 +10,15 @@ import { CartService } from '../cart.service';
 })
 export class HomeComponent implements OnInit {
   products: any[] = [];
-  notificationVisible = false;
+  cartNotificationVisible = false; // Cart notification flag
+  deleteNotificationVisible = false; // Delete notification flag
+  cartNotLoggedNotificationVisible = false; //Not logged notification flag
+  NotificationMessage = '';
 
   constructor(
     private productService: ProductService,
-    private cartService: CartService
+    private cartService: CartService,
+    private authService: AuthService 
   ) {}
 
   ngOnInit(): void {
@@ -40,16 +45,35 @@ export class HomeComponent implements OnInit {
   }
 
   addToCart(product: any): void {
-    this.cartService.addToCart(product);
-    this.notificationVisible = true;
+    if (this.authService.isLoggedIn) {  // Check if the user is logged in
+      this.cartService.addToCart(product);
+      this.cartNotificationVisible = true;
 
-    setTimeout(() => {
-      this.notificationVisible = false;
-    }, 1000);
+      setTimeout(() => {
+        this.cartNotificationVisible = false;
+      }, 1000);
+    } else {
+      this.cartNotLoggedNotificationVisible = true;
+      this.NotificationMessage = 'Please log in to add items to your cart.';
+      
+      setTimeout(() => {
+        this.cartNotLoggedNotificationVisible = false;
+      }, 1000);
+    }
   }
 
   // Delete a product by ID
   deleteProduct(id: number): void {
-    this.productService.deleteProduct(id);
+    if (this.authService.isAdmin) {  // Check if the user is an admin
+      this.productService.deleteProduct(id);
+    } else {
+      this.NotificationMessage = 'You are not allowed to delete products.';
+      this.deleteNotificationVisible = true;
+
+      // Hide the notification after 1 seconds
+      setTimeout(() => {
+        this.deleteNotificationVisible = false;
+      }, 1000);
+    }
   }
 }
